+++ Frends.Kafka.Produce/CHANGELOG.md
@@ -5,7 +5,9 @@
 - New parameter Options.Debug.
 ### Changed 
 - Result.Status and Result.Timestamp have been replaced by Result.Data.
+- Changed how partitions are handled when producing message. See Input.Partition description.
 - All but Kerberos SASL settings can be used on Windows platform.
+- Confluent.Kafka updated from version 1.9.3 to 2.4.
 
 ## [1.2.0] - 2023-10-11
 ### Changed